fix(go-api,go-configeditor): 探测耗时按整数渲染（latency_ms 250 而非 250.0）
改了什么：
- internal/api/server.go：ProbeAvailability.DurationMS 由 float64 改为 **int64**。
- internal/api/probes.go：渲染改用 canonical.NewIntValue（原来用 NewFloat）。
- internal/configeditor/probes.go：接缝处由 float64(result.DurationMS) 改为 int64(...)。
- internal/api/corpus_test.go 与 scripts/gen_management_api_corpus.py 的探测桩：
  duration_ms 由 12.5 改为 **250**（整数，与生产一致）；语料随之重生成（188 条）。

为什么改：
- 参照实现的耗时是**整数**：config_editor.py:1665 用
  `duration_ms = int((monotonic() - started) * 1000)`，字段本身也标注为 `duration_ms: int`
  （config_editor.py:80）。因此 Python 永远输出 `"latency_ms": 250`。
- Go 侧把这个 int 在接缝类型里降级成了 float64，又用 NewFloat 渲染，于是输出
  `"latency_ms": 250.0` —— **每一条真实探测响应都不一致**（不只是边界值）。
- **为什么之前没被发现**：语料桩用的是 `duration_ms=12.5`（一个生产里不可能出现的浮点值），
  于是 Go 与 Python 在 12.5 上"一致"，把这个 bug 完全遮住了。桩不真实，语料就测不到真相。
  这也是本次要同时改桩的原因——只改代码不改桩，语料仍然看不见。

如何验证：
- 修复后：go test ./internal/api/ 通过；gen_management_api_corpus.py --check 退出 0（188 条）。
- **修复前先复现**：只改代码不改桩时测试 FAIL（content-length 259 vs 260），确认桩与语料是
  这个 bug 的唯一遮挡。
- **变异验证（证明语料现在有牙）**：把 probes.go 改回 `NewFloat(float64(...))`，测试立刻 FAIL，
  且失败输出逐字对比出问题：
    got : "latency_ms":250.0
    want: "latency_ms":250
  还原后 CR=0 且全绿。
- 语料里的 latency_ms 取值已从 12.5 变为 250。

顺带说明：该字段原先在接缝类型里丢失 int/float 之别，根因在**类型**而非渲染，所以本次改的是
类型（float64 → int64），没有用"整数就渲染成 int"这类按值猜测的补丁。

+++ internal/api/server.go
@@ -151,9 +151,13 @@ type Server struct {
 
 // ProbeAvailability 是一条探测结果，对应 config_editor 的可用性探测返回值。
 type ProbeAvailability struct {
-	Available  bool
-	URL        string
-	DurationMS float64
+	Available bool
+	URL       string
+	// DurationMS 是**整数**毫秒：参照实现的 config_editor.py:1665 用
+	// `int((monotonic() - started) * 1000)` 取值，字段本身也标注为 `duration_ms: int`
+	// （config_editor.py:80）。因此必须渲染成 JSON 整数——用 float 会得到 `250.0`
+	// 而参照实现是 `250`。这个 int/float 之别原先在接缝类型里丢失了。
+	DurationMS int64
 	Error      string
 }
